JS:s scene-info 분리
